+++ src/proposal/proposal.service.ts
@@ -14,13 +14,13 @@ export class ProposalService {
     private readonly proposalRepository: Repository<ProposalEntity>,
   ) {}
 
-  async findAll(): Promise<ProposalEntity[]> {
+  async findAllPropuesta(): Promise<ProposalEntity[]> {
     return await this.proposalRepository.find({
       relations: ['project', 'professor'],
     });
   }
 
-  async findOne(id: string): Promise<ProposalEntity> {
+  async findPropuestaById(id: string): Promise<ProposalEntity> {
     const proposal: ProposalEntity = await this.proposalRepository.findOne({
       where: { id },
       relations: ['project', 'professor'],
@@ -35,6 +35,12 @@ export class ProposalService {
   }
 
   async create(proposal: ProposalEntity): Promise<ProposalEntity> {
+    if (!proposal.title) {
+      throw new BusinessLogicException(
+        'The proposal title is required',
+        BusinessError.INVALID_DATA,
+      );
+    }
     return await this.proposalRepository.save(proposal);
   }
 
@@ -55,7 +61,7 @@ export class ProposalService {
     });
   }
 
-  async delete(id: string) {
+  async deletePropuesta(id: string) {
     const proposal: ProposalEntity = await this.proposalRepository.findOne({
       where: { id },
     });
@@ -64,6 +70,13 @@ export class ProposalService {
         'The proposal with the given id was not found',
         BusinessError.NOT_FOUND,
       );
+
+    if (proposal.project) {
+      throw new BusinessLogicException(
+        'The proposal has a project assigned',
+        BusinessError.INVALID_DATA,
+      );
+    }
 
     await this.proposalRepository.remove(proposal);
   }
